crewai_version/requirements.txt
streamlit==1.33.0
python-dotenv==1.0.1
openai==1.58.1
pydantic==2.8.2
langchain-openai==0.3.0
crewai==0.98.0
httpx==0.27.2
